requirements.txt
flask
python-dotenv
groq
scikit-learn
pandas
joblib
gunicorn
